+++ ZGBijouRoller.lua
@@ -33,6 +33,15 @@ local ROLL = {
   GREED = 2
 }
 
+local function getRollText(index)
+  for k, v in pairs(ROLL) do
+    if (v == index) then
+      return k
+    end
+  end
+  return "Missing Roll in getRollText(index)"
+end
+
 local rollOnCoin = {}
 local rollOnBijou = {}
 
@@ -62,32 +71,158 @@ end
 
 ---- options
 
+local rollDefault = ROLL.GREED
 local debug = false
-local rollBijou = ROLL.GREED
-local seperateRolls = false
-local rollCoin = rollBijou
 
 --------------------------------------------------------------
 ---- helper functions
 --------------------------------------------------------------
 
-local EVENT_NAME = "ZGBijouRoller"
+local addonName = "ZGBijouRoller"
 
-local bijouFrame = CreateFrame("Frame", EVENT_NAME)
+local bijouFrame = CreateFrame("Frame", addonName)
+local panel = CreateFrame("Frame", addonName .. "Panel")
+local saveSettingsButton = CreateFrame("CheckButton", "saveSettingsButton", panel, "ChatConfigCheckButtonTemplate")
+saveSettingsButton:SetScript("OnClick",
+  function()
+    ZGBijouRollerSettings["saveSettings"] = not ZGBijouRollerSettings["saveSettings"]
+  end
+)
+local bijouDropDown = CreateFrame("FRAME", "bijouDropDown", panel, "UIDropDownMenuTemplate")
+local coinDropDown = CreateFrame("FRAME", "coinDropDown", panel, "UIDropDownMenuTemplate")
+local seperateRollsButton = CreateFrame("CheckButton", "seperateRollsButton", panel, "ChatConfigCheckButtonTemplate")
+seperateRollsButton:SetScript("OnClick",
+  function()
+    ZGBijouRollerSettings["seperateRolls"] = not ZGBijouRollerSettings["seperateRolls"]
+    if (ZGBijouRollerSettings["seperateRolls"]) then
+      coinDropDown:Show()
+    else
+      coinDropDown:Hide()
+    end
+  end
+)
 
 -- print the current roll behaviour
 local function printRollBehavior()
   for k, v in pairs(ROLL) do
-    if (v == rollBijou) then
+    if (v == ZGBijouRollerSettings["rollBijou"]) then
       print(format("current roll behaviour for BIJOUs: %s", k))
     end
-    if (seperateRolls and v == rollCoin) then
+    if (ZGBijouRollerSettings["seperateRolls"] and v == ZGBijouRollerSettings["rollCoin"]) then
       print(format("current roll behaviour for COINs:  %s", k))
     end
   end
-  if (not seperateRolls) then
+  if (not ZGBijouRollerSettings["seperateRolls"]) then
     print("coin roll behaviour is locked to bijou behaviour")
   end
+end
+
+-- function for clicking on the bijou dropdown in the options panel
+local function bijouDropDown_OnClick(_, arg1)
+  ZGBijouRollerSettings["rollBijou"] = arg1
+  UIDropDownMenu_SetText(bijouDropDown, getRollText(arg1))
+end
+
+-- initialization of the bojou dropdown panel
+local function initBijouDropDown(dropDown, level, menuList)
+  local info = UIDropDownMenu_CreateInfo()
+  info.func = bijouDropDown_OnClick
+  for k, v in pairs(ROLL) do
+    info.text = k
+    info.arg1 = v
+    info.checked = v == ZGBijouRollerSettings["rollBijou"]
+    UIDropDownMenu_AddButton(info)
+  end
+end
+
+-- function for clicking on the coin dropdown in the options panel
+local function coinDropDown_OnClick(_, arg1)
+  ZGBijouRollerSettings["rollCoin"] = arg1
+  UIDropDownMenu_SetText(coinDropDown, getRollText(arg1))
+end
+
+-- initialization of the coin dropdown panel
+local function initCoinDropDown(dropDown, level, menuList)
+  local info = UIDropDownMenu_CreateInfo()
+  info.func = coinDropDown_OnClick
+  for k, v in pairs(ROLL) do
+    info.text = k
+    info.arg1 = v
+    info.checked = v == ZGBijouRollerSettings["rollCoin"]
+    UIDropDownMenu_AddButton(info)
+  end
+end
+
+-- setter for ZGBijouRollerSettings["seperateRolls"]
+local function setSeperateRolls(value)
+  ZGBijouRollerSettings["seperateRolls"] = value
+  seperateRollsButton:SetChecked(ZGBijouRollerSettings["seperateRolls"])
+  if (ZGBijouRollerSettings["seperateRolls"]) then
+    coinDropDown:Show()
+  else
+    coinDropDown:Hide()
+  end
+end
+
+-- initialization of the addon panel
+local function InitializePanel()
+  panel.name = "ZG Bijou Roller"
+
+  InterfaceOptions_AddCategory(panel)
+
+  local title = panel:CreateFontString(addonName .. "Title", "OVERLAY", "GameFontNormalLarge")
+  title:SetPoint("TOP", 0, -12)
+  title:SetText(panel.name)
+
+  -- save settings
+  local saveSettingsText = panel:CreateFontString(addonName .. "saveSettingsText", "OVERLAY", "GameFontNormalSmall")
+  saveSettingsText:SetPoint("TOPLEFT", 35, -45)
+  saveSettingsText:SetText("Load all these settings after a relog / reloadui ?")
+
+  local saveSettingsText2 = panel:CreateFontString(addonName .. "saveSettingsText2", "OVERLAY", "GameFontNormalSmall")
+  saveSettingsText2:SetPoint("TOPLEFT", 35, -65)
+  saveSettingsText2:SetText(format("If not, the default roll behaviour (%s) will ge selected again.", getRollText(rollDefault)))
+
+  saveSettingsButton:SetPoint("TOPLEFT", 10, -48)
+  saveSettingsButton:SetChecked(ZGBijouRollerSettings["saveSettings"])
+
+  -- bijou behaviour
+  local bijouRollText = panel:CreateFontString(addonName .. "bijouRollText", "OVERLAY", "GameFontNormalSmall")
+  bijouRollText:SetPoint("TOPLEFT", 10, -135)
+  bijouRollText:SetText("Bijou rolls:")
+
+  bijouDropDown:SetPoint("TOPLEFT", 0, -150)
+  UIDropDownMenu_SetText(bijouDropDown, getRollText(ZGBijouRollerSettings["rollBijou"]))
+  UIDropDownMenu_Initialize(bijouDropDown, initBijouDropDown, 1)
+
+  local seperateRollsText = panel:CreateFontString(addonName .. "seperateRollsText", "OVERLAY", "GameFontNormalSmall")
+  seperateRollsText:SetPoint("TOPLEFT", 220, -135)
+  seperateRollsText:SetText("Roll Coins separably?")
+
+  seperateRollsButton:SetPoint("TOPLEFT", 260, -150)
+  seperateRollsButton:SetChecked(ZGBijouRollerSettings["seperateRolls"])
+
+  -- coin behaviour
+  local coinRollText = panel:CreateFontString(addonName .. "coinRollText", "OVERLAY", "GameFontNormalSmall")
+  coinRollText:SetPoint("TOPLEFT", 360, -135)
+  coinRollText:SetText("Coin rolls:")
+
+  coinDropDown:SetPoint("TOPLEFT", 350, -150)
+  UIDropDownMenu_SetText(coinDropDown, getRollText(ZGBijouRollerSettings["rollCoin"]))
+  UIDropDownMenu_Initialize(coinDropDown, initCoinDropDown, 1)
+
+  if (not ZGBijouRollerSettings["seperateRolls"]) then
+    coinDropDown:Hide()
+  end
+
+  -- info footer
+  local info1 = panel:CreateFontString(addonName .. "info1", "OVERLAY", "GameFontNormalSmall")
+  info1:SetPoint("TOPLEFT", 70, -335)
+  info1:SetText("All options will be saved immediatly without using the Okay or Cancel Button below.")
+
+  local helpText = panel:CreateFontString(addonName .. "Help", "OVERLAY", "GameFontNormalSmall")
+  helpText:SetPoint("TOPLEFT", 10, -400)
+  helpText:SetText("For more information, type /zgroll help")
 end
 
 -- ADDON_LOADED
@@ -95,7 +230,20 @@ function bijouFrame:ADDON_LOADED(eventName)
   if (debug) then
     print(format("event found in bijouFrame:ADDON_LOADED: %s", eventName))
   end
-  if (eventName == EVENT_NAME) then
+  if (eventName == addonName) then
+    if (ZGBijouRollerSettings == nil or not ZGBijouRollerSettings["saveSettings"]) then
+      ZGBijouRollerSettings = {}
+      ZGBijouRollerSettings["rollBijou"] = rollDefault
+      ZGBijouRollerSettings["rollCoin"] = rollDefault
+    else
+      if (ZGBijouRollerSettings["rollBijou"] == nil) then
+        ZGBijouRollerSettings["rollBijou"] = rollDefault
+      end
+      if (ZGBijouRollerSettings["rollCoin"] == nil) then
+        ZGBijouRollerSettings["rollCoin"] = rollDefault
+      end
+    end
+    InitializePanel()
     for k, _ in pairs(COIN_IDS) do
       rollOnCoin:addItem(COIN_IDS[k])
       if (debug) then
@@ -124,30 +272,30 @@ function bijouFrame:START_LOOT_ROLL(rollID)
       print(format("canNeed: %s", tostring(canNeed)))
       print(format("canGreed: %s", tostring(canGreed)))
     end
-    if (seperateRolls and rollOnCoin[name]
-      and (rollCoin == ROLL.NEED and canNeed
-        or rollCoin == ROLL.GREED and canGreed
-        or rollCoin == ROLL.PASS)) then
+    if (ZGBijouRollerSettings["seperateRolls"] and rollOnCoin[name]
+      and (ZGBijouRollerSettings["rollCoin"] == ROLL.NEED and canNeed
+        or ZGBijouRollerSettings["rollCoin"] == ROLL.GREED and canGreed
+        or ZGBijouRollerSettings["rollCoin"] == ROLL.PASS)) then
       if (debug) then
-        print(format("item match. rolling for coin: %s", rollCoin))
+        print(format("item match. rolling for coin: %s", ZGBijouRollerSettings["rollCoin"]))
       end
-      RollOnLoot(rollID, rollCoin)
-    elseif (rollBijou == ROLL.NEED and canNeed
-      or rollBijou == ROLL.GREED and canGreed
-      or rollBijou == ROLL.PASS) then
+      RollOnLoot(rollID, ZGBijouRollerSettings["rollCoin"])
+    elseif (ZGBijouRollerSettings["rollBijou"] == ROLL.NEED and canNeed
+      or ZGBijouRollerSettings["rollBijou"] == ROLL.GREED and canGreed
+      or ZGBijouRollerSettings["rollBijou"] == ROLL.PASS) then
       if (debug) then
-        print(format("item match. rolling for bijou: %s", rollBijou))
+        print(format("item match. rolling for bijou: %s", ZGBijouRollerSettings["rollBijou"]))
       end
-      RollOnLoot(rollID, rollBijou)
+      RollOnLoot(rollID, ZGBijouRollerSettings["rollBijou"])
     else
       print("You found a bug, please report it with the following informations!")
       print("Buginfo:")
       print(string.format("name: %s", name))
       print(string.format("canNeed: %s", canNeed))
       print(string.format("canGreed: %s", canGreed))
-      print(string.format("rollCoin: %s", rollCoin))
-      print(string.format("rollBijou: %s", rollBijou))
-      print(string.format("seperateRolls: %s", seperateRolls))
+      print(string.format("rollCoin: %s", ZGBijouRollerSettings["rollCoin"]))
+      print(string.format("rollBijou: %s", ZGBijouRollerSettings["rollBijou"]))
+      print(string.format("seperateRolls: %s", ZGBijouRollerSettings["seperateRolls"]))
     end
   end
 end
@@ -185,35 +333,38 @@ end
 -- handler for bijou commands
 local function handlerBijou(msg)
   if (msg == "need") then
-    rollBijou = ROLL.NEED
+    bijouDropDown_OnClick(nil, ROLL.NEED)
     printRollBehavior()
   elseif (msg == "greed") then
-    rollBijou = ROLL.GREED
+    bijouDropDown_OnClick(nil, ROLL.GREED)
     printRollBehavior()
   elseif (msg == "pass") then
-    rollBijou = ROLL.PASS
+    bijouDropDown_OnClick(nil, ROLL.PASS)
     printRollBehavior()
-  else
+  elseif (msg == "help") then
     printHelp()
+  else
+    InterfaceOptionsFrame_OpenToCategory(panel)
+    InterfaceOptionsFrame_OpenToCategory(panel) -- one call will only open the normal interface options. blizzard bug?
   end
 end
 
 -- handler for coin commands
 local function handlerCoin(msg)
   if (msg == "need") then
-    rollCoin = ROLL.NEED
-    seperateRolls = true
+    coinDropDown_OnClick(nil, ROLL.NEED)
+    setSeperateRolls(true)
     printRollBehavior()
   elseif (msg == "greed") then
-    rollCoin = ROLL.GREED
-    seperateRolls = true
+    coinDropDown_OnClick(nil, ROLL.GREED)
+    setSeperateRolls(true)
     printRollBehavior()
   elseif (msg == "pass") then
-    rollCoin = ROLL.PASS
-    seperateRolls = true
+    coinDropDown_OnClick(nil, ROLL.PASS)
+    setSeperateRolls(true)
     printRollBehavior()
   elseif (msg == "lock") then
-    seperateRolls = false
+    setSeperateRolls(false)
     printRollBehavior()
   else
     printHelp()